+++ Sources/App/ViewModel/ThermostatViewModelable.swift
@@ -13,16 +13,16 @@ enum HeatingCoolingState: Int { case OFF = 0, HEAT, COOL, AUTO }
 protocol ThermostatViewModelable
 {
     //MARK: Subscribers
-    var currentOutdoorTemperatureObserver : PublishSubject<Int> {get}
-    var currentIndoorHumidityObserver : PublishSubject<Int> {get}
-    var currentIndoorTemperatureObserver : PublishSubject<Int> {get}
-    var targetIndoorTemperatureObserver : PublishSubject<Int> {get}
+    var currentOutdoorTemperatureObserver : PublishSubject<Double> {get}
+    var currentIndoorHumidityObserver : PublishSubject<Double> {get}
+    var currentIndoorTemperatureObserver : PublishSubject<Double> {get}
+    var targetIndoorTemperatureObserver : PublishSubject<Double> {get}
     var currentHeatingCoolingStateObserver : PublishSubject<HeatingCoolingState> {get}
     var targetHeatingCoolingStateObserver : PublishSubject<HeatingCoolingState> {get}
     var forcingWaterHeaterObserver : PublishSubject<Int> {get}
-    var boilerHeatingLevelObserver : PublishSubject<Int> {get}
+    var boilerHeatingLevelObserver : PublishSubject<Double> {get}
     //MARK: Actions
-    var targetTemperaturePublisher : PublishSubject<Int> {get}
+    var targetTemperaturePublisher : PublishSubject<Double> {get}
     var targetHeatingCoolingStatePublisher : PublishSubject<HeatingCoolingState> {get}
     var forcingWaterHeaterPublisher : PublishSubject<Int> {get}
     //MARK: Dispatcher
